FIXED:Animation got sent as CellID FIXED: Various other bugs
git-svn-id: https://svn.code.sf.net/p/oo-open/code@462 2a060ecb-2c39-0410-a787-992002912dfc

+++ trunk/OblivionOnlineServer/PositionChunkHandler.cpp
@@ -57,6 +57,6 @@ size_t HandleAnimationChunk(GameServer *gs,InPacket *pkg, BYTE* chunkdata,size_t
 		return 0;
 	}	
 	ent->SetAnimation(AnimationID,AnimStatus);
-	gs->GetIO()<<LogLevel::SystemMessage<<"Animation State changed Entity : " << FormID<<" Status "<< status<< "animation "<< AnimationID << " On /Off " << AnimStatus;  
+	gs->GetIO()<<LogLevel::SystemMessage<<"Animation State changed Entity : " << FormID<<" Status "<< (unsigned short)status<< "animation "<< AnimationID << " On /Off " << AnimStatus;  
 	return GetMinChunkSize(PkgChunk::Animation);
 }
